Add v2 API
The v2 API is a new API which provides a superset of the
old API functionality aiming at providing support for
different protocols, and more features.
Including, between others the following additions:
  * New CCA (clear channel assesment) procedure support
  * In Tx:
    * Transmissions truncating part of the preamble
    * Transmitters emiting interference before and/or after
      the actual packet
  * In Rx:
    * Multi address Rx support
    * Piggybacking reception procedures/payloads
    * Opening the receiver after the preamble has started
    * Receivers staying in Rx shorter/longer than the actual Tx packet
      (to simulate lenght or coding rate errors)
  * Both for Tx & Rx:
    * Up to 64 bit address support (from 32)
    * Different FEC coding rates (with the same  modulation)

The old API is NOT deprecated, and remains both API and ABI compatible.
Users are at this point encouraged to NOT switch to the new API
at this point. The commitment on ABI compatibility remains in place.

This new APIv2 should be considered in Alpha state.
Changes to this API should be expected.

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>