Rename enum element in unused API
One element name in an enum was very likely to cause conflicts
with other enums.
Even though this could be considered an API, it is a
not used API, and therefore there should be no impact.

If as a user you have hit a problem here, please raise an
issue to the maintainer.

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>